Extract mapping logic
