+++ source/LearningTests/Infrastructure/Transactions/Aggregations/AggregationResultOverHttpTests.cs
@@ -13,6 +13,7 @@
 // limitations under the License.
 
 using System.Reflection;
+using Messaging.Infrastructure.Configuration.Serialization;
 using Messaging.Infrastructure.Transactions.Aggregations;
 using Microsoft.Extensions.Configuration;
 
@@ -21,19 +22,22 @@ namespace LearningTests.Infrastructure.Transactions.Aggregations;
 public class AggregationResultOverHttpTests
 {
     private readonly IConfigurationRoot _configuration;
+    private readonly AggregationResultMapper _mapper;
 
     public AggregationResultOverHttpTests()
     {
         _configuration = new ConfigurationBuilder()
             .AddUserSecrets(Assembly.GetExecutingAssembly(), false, false)
             .Build();
+
+        _mapper = new AggregationResultMapper(new Serializer());
     }
 
     [Fact]
     public async Task Service_is_unavailable()
     {
         using var httpClient = new HttpClient();
-        var service = new AggregationResultsOverHttp(httpClient, new Uri($"{_configuration["ServiceUri"]!}/WrongUri"));
+        var service = new AggregationResultsOverHttp(httpClient, new Uri($"{_configuration["ServiceUri"]!}/WrongUri"), _mapper);
 
         await Assert.ThrowsAnyAsync<Exception>(() => service.GetResultAsync(Guid.Parse(_configuration["BatchId"]!), _configuration["GridArea"]!)).ConfigureAwait(false);
     }
@@ -42,7 +46,7 @@ public class AggregationResultOverHttpTests
     public async Task Can_retrieve_result()
     {
         using var httpClient = new HttpClient();
-        var service = new AggregationResultsOverHttp(httpClient, new Uri(_configuration["ServiceUri"]!));
+        var service = new AggregationResultsOverHttp(httpClient, new Uri(_configuration["ServiceUri"]!), _mapper);
 
         var result = await service.GetResultAsync(Guid.Parse(_configuration["BatchId"]!), _configuration["GridArea"]!).ConfigureAwait(false);
